requirements.txt
attrs==22.1.0
aws-cdk-lib==2.35.0
boto3==1.11.14
botocore==1.14.17
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
constructs==10.1.141
distlib==0.3.6
#docker==6.0.1
docutils==0.15.2
exceptiongroup==1.0.0
filelock==3.9.0
idna==3.4
jmespath==0.10.0
jsii==1.70.0
lxml==4.9.1
packaging==23.0
Pillow==9.3.0
pipenv==2023.2.18
platformdirs==3.0.0
protobuf==4.21.12
publication==0.0.3
python-dateutil==2.8.2
python-pptx==0.6.18
requests==2.28.2
s3transfer==0.3.7
six==1.16.0
tk==0.1.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.25.11
virtualenv==20.19.0
virtualenv-clone==0.5.7
websocket-client==1.5.1
XlsxWriter==3.0.3
